infra/google-budget-function/requirements.txt
functions-framework==3.*
google-cloud-service-usage==1.*
requests==2.*
